refactor: reorganize model folder structure
- Merge models/towers/ and models/towers_v2/ into models/towers/ — move
  cost_class.dart into the active folder and drop the legacy v1 shell
- Move models/base_model.dart into models/base/ alongside the other base models
- Move models/maps/map.dart into models/maps/map/ to match the nested
  pattern used by bloons and boss models
- Update all imports across lib/ and test/
