+++ lib/models/base/base_bloon.dart
@@ -1,4 +1,4 @@
-import '/models/base_model.dart';
+import '/models/base/base_model.dart';
 
 class BaseBloon extends BaseModel {
   final bool isMoab;
