+++ lib/models/towers/common/upgrade_class.dart
