+++ lib/presentation/screens/hero/hero_skins.dart
@@ -3,8 +3,8 @@ import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
-import '/models/towers_v2/hero/hero.dart';
-import '/models/towers_v2/hero/hero_skin.dart';
+import '/models/towers/hero/hero.dart';
+import '/models/towers/hero/hero_skin.dart';
 import '/presentation/widgets/common/app_image.dart';
 import '/presentation/widgets/common/detail_page_scaffold.dart';
 import '/presentation/widgets/common/image_carousel.dart';
