+++ lib/presentation/screens/hero/single_hero.dart
@@ -5,8 +5,8 @@ import 'package:flutter/services.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
-import '/models/towers_v2/hero/hero.dart';
-import '/models/towers_v2/common/hero_level_class.dart';
+import '/models/towers/hero/hero.dart';
+import '/models/towers/common/hero_level_class.dart';
 import '/presentation/widgets/common/detail_page_scaffold.dart';
 import '/presentation/widgets/common/loader.dart';
 import '/presentation/widgets/common/image_carousel.dart';
