+++ lib/presentation/widgets/heroes/hero_level.dart
@@ -1,6 +1,6 @@
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
-import '/models/towers_v2/common/hero_level_class.dart';
+import '/models/towers/common/hero_level_class.dart';
 import '/presentation/widgets/common/app_image.dart';
 import '/presentation/widgets/common/carousel_with_indicator.dart';
 import '/presentation/widgets/common/stat_row.dart';
